feat: implement token-based sessions and automatic player cleanup

+++ src/main/java/com/poker/model/PlayerAction.java
@@ -1,3 +1,3 @@
 package com.poker.model;
 
-public record PlayerAction(ActionType type, int amount) {}
+public record PlayerAction(ActionType type, long amount) {}
